+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 # Xcodeproj Changelog
 
-## master
+## Master
+
+* `PlistHelper`: Now the `plutil` tool is used to save the files if available to
+  produce output consistent with Xcode.  
+  [Fabio Pelosin](https://github.com/irrationalfab)
+
+* `PBXGroup`: Now file references to Xcode projects are properly handled and
+  setup. Also the `ObjectDictionary` class has been improved and now can be
+  used to edit the attributes using it.
+  [Fabio Pelosin](https://github.com/irrationalfab)
+  [Xcodeproj#172](https://github.com/CocoaPods/Xcodeproj/pull/172)
+
+## 0.18.0
+
+###### Enhancements
+
+* [PlistHelper] The native extension has been removed in favour of the usage of
+  the `plutil` tool to read ASCII property list files.  
+  [Fabio Pelosin](https://github.com/irrationalfab)
+  [Xcodeproj#168](https://github.com/CocoaPods/Xcodeproj/pull/168)
+  [Xcodeproj#167](https://github.com/CocoaPods/Xcodeproj/issues/167)
+
+* [PBXFileReference] If a file reference represents an external Xcode project
+  and is removed from the project then all items related to the external
+  project will also be removed.  
+  [JP Simard](https://github.com/jpsim)
+  [Eloy Durán](https://github.com/alloy)
+  [Xcodeproj#158](https://github.com/CocoaPods/Xcodeproj/issues/158)
+  [Xcodeproj#161](https://github.com/CocoaPods/Xcodeproj/pull/161)
 
 ###### Bug fixes
 
@@ -13,8 +41,8 @@
 ###### Enhancements
 
 * [Workspace] Added support for file references.  
-  [Kyle Fuller][kylef]
-  [Fabio Pelosin][irrationalfab]
+  [Kyle Fuller](https://github.com/kylef)
+  [Fabio Pelosin](https://github.com/irrationalfab)
   [Xcodeproj#105](https://github.com/CocoaPods/Xcodeproj/pull/150)
 
 ## 0.16.1
@@ -77,7 +105,7 @@
 
 ###### Bug Fixes
 
-* [Extension] Fixed intermittent `Xcodeproj.write_plist` error.  
+* [Extension] Fixed intermittent `Xcodeproj::PlistHelper.write_plist` error.  
   [Eloy Durán](https://github.com/alloy)
 
 
@@ -399,6 +427,3 @@
 
 * The file type of the frameworks file references has be corrected.
 
-
-[irrationalfab]: https://github.com/irrationalfab
-[kylef]: (https://github.com/kylef)